refactor(tests): align service test files with 1:1 module naming convention
- Merge user service test files into test_user.py
- Rename test_summary_module.py to test_summary.py
- Merge test_coordination_increment.py into test_coordination.py
- Remove 7 non-conforming entries from ALLOWLIST_NON_CONFORMING_TESTS in test_init.py
