+++ tests/server/services/processor_modules/test_summary.py
