--- tests/server/services/test_coordination_increment.py
@@ -1,40 +1,0 @@
-from supernote.server.services.coordination import (
-    SqliteCoordinationService,
-)
-
-
-async def test_increment(coordination_service: SqliteCoordinationService) -> None:
-    key = "incr:test"
-
-    # 1. Increment new key
-    val = await coordination_service.increment(key, 1, ttl=60)
-    assert val == 1
-
-    # Check it exists
-    stored = await coordination_service.get_value(key)
-    assert stored == "1"
-
-    # 2. Increment existing
-    val = await coordination_service.increment(key, 1)
-    assert val == 2
-
-    # 3. Increment by larger amount
-    val = await coordination_service.increment(key, 10)
-    assert val == 12
-
-
-async def test_increment_expiry(
-    coordination_service: SqliteCoordinationService,
-) -> None:
-    key = "incr:expire"
-
-    # Set with short TTL (but we can't easily wait for it in unit test without sleep)
-    # Instead, we manually set an expired value first?
-    # No, let's use the property that increment sets TTL on creation.
-
-    # 1. Create
-    val = await coordination_service.increment(key, 1, ttl=100)
-    assert val == 1
-
-    # 2. Verify TTL is roughly logic (hard to verify exact value without inspecting DB directly)
-    # But we can verify "expired" behavior by mocking time or inserting expired row manually.
