--- tests/server/services/test_user_password.py
@@ -1,52 +1,0 @@
-import hashlib
-
-import pytest
-from sqlalchemy import select
-
-from supernote.server.config import ServerConfig
-from supernote.server.db.models.user import UserDO
-from supernote.server.db.session import DatabaseSessionManager
-from supernote.server.services.coordination import CoordinationService
-from supernote.server.services.user import UserService
-
-
-async def test_retrieve_password_success(
-    user_service: UserService,
-    session_manager: DatabaseSessionManager,
-    coordination_service: CoordinationService,
-    server_config: ServerConfig,
-) -> None:
-    # Setup
-    email = "forgot@example.com"
-    old_pw = hashlib.md5(b"old").hexdigest()
-    new_pw = hashlib.md5(b"new").hexdigest()
-
-    # Create user directly in DB
-    async with session_manager.session() as session:
-        user = UserDO(
-            email=email, password_md5=old_pw, display_name="Forgot", is_active=True
-        )
-        session.add(user)
-        await session.commit()
-
-    # Validation: Pass scalars
-    result: bool = await user_service.retrieve_password(email, new_pw)
-    assert result is True
-
-    # Verify in DB
-    async with session_manager.session() as session:
-        user_result = await session.execute(select(UserDO).where(UserDO.email == email))
-        user = user_result.scalar_one()
-        assert user.password_md5 == new_pw
-
-
-async def test_retrieve_password_invalid_md5(user_service: UserService) -> None:
-    with pytest.raises(ValueError, match="Invalid password format"):
-        await user_service.retrieve_password("user@example.com", "plain_text_password")
-
-
-async def test_retrieve_password_user_not_found(user_service: UserService) -> None:
-    result = await user_service.retrieve_password(
-        "nonexistent@example.com", hashlib.md5(b"pw").hexdigest()
-    )
-    assert result is False
