+++ tests/test_init.py
@@ -55,15 +55,8 @@ ALLOWLIST_NON_CONFORMING_TESTS = {
     "server/routes/test_oss_upload_security_extra.py",
     "server/routes/test_system_extended.py",
     "server/services/processor_modules/test_processor_integration.py",
-    "server/services/processor_modules/test_summary_module.py",
     "server/services/test_blob_integration.py",
-    "server/services/test_coordination_increment.py",
     "server/services/test_processor_modules.py",
-    "server/services/test_user_bootstrap.py",
-    "server/services/test_user_password.py",
-    "server/services/test_user_refactor.py",
-    "server/services/test_user_token_expiration.py",
-    "server/services/test_user_validation.py",
     "server/test_trace.py",
     "server/web/test_capacity.py",
     "server/web/test_device_vs_web_structure.py",
